requirements.txt
openai
click
python-dotenv
jsonschema